make imageOptionalHeader32 class & check the whole code

+++ targetFile.h
@@ -8,20 +8,19 @@
 
 class TargetFile {
     std::byte * whole_bin_ = nullptr;
-
-    bool     is_32bit_ = false;
-    unsigned sz_       = 0;
+    bool        is_32bit_  = false;
+    size_t      sz_        = 0;
 
 public:
-    static void     getSubBytes(std::byte *, const std::byte *, unsigned, unsigned);
-    static unsigned getSubBytes(const std::byte *, unsigned, unsigned);
+    static void   getSubBytes(std::byte *, const std::byte *, size_t, size_t);
+    static size_t getSubBytes(const std::byte *, size_t, size_t);
 
     explicit TargetFile(const std::string &);
     // TargetFile(TargetFile &);
     ~TargetFile();
 
-                  void     getFileContents(std::byte *, unsigned, unsigned) const;
-    [[nodiscard]] unsigned getFileContents(unsigned, unsigned) const;
+                  void     getFileContents(std::byte *, size_t, size_t) const;
+    [[nodiscard]] unsigned getFileContents(size_t, size_t) const;
     [[nodiscard]] bool     getIs32bit() const;
 };
